Experiments
Toying with the use of import/export, dates in objects, and
getting comfortable using the new "dates-fn" npm package.
Still thinking through the logic of the application
a little bit as well. Getting every task method from
its module could be a good way to keep the logic separated
from DOM related things, so I think that's the way I'll
continue for now.